site-(site-admin)-文章列表增加删除文章确定

+++ site/site-admin/src/views/articleList/index.vue
@@ -1,11 +1,11 @@
 <template>
 	<div class="article-list-container">
 		<div class="add-article">
-			<router-link :to="{name:'ArticleEdit'}">
+			<router-link :to="{ name: 'ArticleEdit' }">
 				<el-button type="success">新增文章</el-button>
 			</router-link>
 		</div>
-    <!-- 表格组件 -->
+		<!-- 表格组件 -->
 		<div class="article-table-container">
 			<el-table
 				:data="articles.rows"
@@ -116,24 +116,35 @@
 					header-align="center"
 				>
 					<template slot-scope="scope">
-						<el-button size="mini" type="primary" @click="onHandleEdit(scope)">编辑</el-button>
-						<el-button size="mini" type="danger" @click="onHandleDelete(scope)">删除</el-button>
+						<el-button
+							size="mini"
+							type="primary"
+							@click="onHandleEdit(scope)"
+							>编辑</el-button
+						>
+						<el-button
+							size="mini"
+							type="danger"
+							@click="onHandleDelete(scope)"
+							>删除</el-button
+						>
 					</template>
 				</el-table-column>
 			</el-table>
 		</div>
-    <!-- 分页组件 -->
-    <div class="page-container">
-    <el-pagination
-      @size-change="handleSizeChange"
-      @current-change="handleCurrentChange"
-      :current-page="page"
-      :page-sizes="[5,10, 20, 30, 50]"
-      :page-size="limit"
-      layout="total, sizes, prev, pager, next, jumper"
-      :total="articles.total">
-    </el-pagination>
-    </div>
+		<!-- 分页组件 -->
+		<div class="page-container">
+			<el-pagination
+				@size-change="handleSizeChange"
+				@current-change="handleCurrentChange"
+				:current-page="page"
+				:page-sizes="[5, 10, 20, 30, 50]"
+				:page-size="limit"
+				layout="total, sizes, prev, pager, next, jumper"
+				:total="articles.total"
+			>
+			</el-pagination>
+		</div>
 	</div>
 </template>
 
@@ -147,7 +158,7 @@ export default {
 			articles: [], //文章
 			page: 1, //当前的页码
 			limit: 10, //每页的数据
-      myBaseUrl,//图片的基础路径
+			myBaseUrl, //图片的基础路径
 		};
 	},
 	created() {
@@ -159,35 +170,50 @@ export default {
 			this.articles = await articleApi.getArticle(this.page, this.limit);
 		},
 		formateDate,
-    // 控制每页多少条数据
-    handleSizeChange(val){
-      this.limit = val;
-      // 重新获取数据
-      this.__initArticleData();
-    },
-    // 当前第多少也页
-    handleCurrentChange(val){
-      this.page = val;
-      // 重新获取数据
-      this.__initArticleData();
-    },
-    // 修改文章
-    async onHandleEdit(scope){
-      const data = scope.row;//文章数据
-      const articleId = scope.row.id;//文章ID
-      this.$router.push({name:'ArticleEdit',query:{data:data,articleId}})
-    },
-    // 删除文章
-    async onHandleDelete(scope) {
-      const articleId = scope.row.id;//文章ID
-      const res = await articleApi.deleteArticle(articleId);//删除
-      this.$message({
-        type:'success',
-        message:res,
-duration:2000
-      });
-	  this.__initArticleData();//重新获取数据
-    }
+		// 控制每页多少条数据
+		handleSizeChange(val) {
+			this.limit = val;
+			// 重新获取数据
+			this.__initArticleData();
+		},
+		// 当前第多少也页
+		handleCurrentChange(val) {
+			this.page = val;
+			// 重新获取数据
+			this.__initArticleData();
+		},
+		// 修改文章
+		async onHandleEdit(scope) {
+			const data = scope.row; //文章数据
+			const articleId = scope.row.id; //文章ID
+			this.$router.push({
+				name: "ArticleEdit",
+				query: { data: data, articleId },
+			});
+		},
+		// 删除文章
+		async onHandleDelete(scope) {
+			this.$confirm("此操作将永久删除该文章, 是否继续?", "提示", {
+				confirmButtonText: "确定",
+				cancelButtonText: "取消",
+				type: "warning",
+			})
+				.then(async () => {
+					const articleId = scope.row.id; //文章ID
+					const res = await articleApi.deleteArticle(articleId); //删除
+					this.$message({
+						type: "success",
+						message: res,
+					});
+          this.__initArticleData(); //重新获取数据
+				})
+				.catch(() => {
+					this.$message({
+						type: "info",
+						message: "已取消删除",
+					});
+				});
+		},
 	},
 };
 </script>
@@ -201,12 +227,12 @@ duration:2000
 	width: 90%;
 	margin: 70px auto;
 }
-.page-container{
-  width: 100%;
-  margin-top: 40px;
-  margin-left: 60px;
+.page-container {
+	width: 100%;
+	margin-top: 40px;
+	margin-left: 60px;
 }
-.add-article{
+.add-article {
 	position: absolute;
 	right: 80px;
 	top: -60px;
